% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorOneWithMetadata.R
\name{anchorOneWithMetadata}
\alias{anchorOneWithMetadata}
\title{Extract the first anchor from GenomicInteractions object
with metadata}
\usage{
anchorOneWithMetadata(gint)
}
\arguments{
\item{gint}{A GenomicInteractions object}
}
\value{
gr a GRanges object with metadata
}
\description{
Creates a GRanges object from the first anchor of a GenomicInteractions
object, but extracts all metadata columns.
}
\examples{
 load(system.file(
 "extdata","gs3dk_sample_data.rda",
 package = "GS3DKViz"))
 anchorOneWithMetadata(gs3dk_gint)
}
\keyword{GenomicInteractions}
\keyword{GenomicRanges}
\keyword{metadata}
