% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorTwoWithMetadata.R
\name{anchorTwoWithMetadata}
\alias{anchorTwoWithMetadata}
\title{Extract the second anchor from GenomicInteractions object
with metadata}
\usage{
anchorTwoWithMetadata(gint)
}
\arguments{
\item{gint}{A GenomicInteractions object}
}
\value{
gr a GRanges object with metadata
}
\description{
Creates a GRanges object from the second anchor of a GenomicInteractions
object, but extracts all metadata columns.
}
\examples{
 load(system.file(
 "extdata","gs3dk_sample_data.rda",
 package = "GS3DKViz"))
 anchorTwoWithMetadata(gs3dk_gint)
}
\keyword{GenomicInteractions}
\keyword{GenomicRanges}
\keyword{metadata}
