% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractionDetail.R
\name{plotInteractionDetail}
\alias{plotInteractionDetail}
\title{Plot a GenomicInteractions object with details for Enhancers}
\usage{
plotInteractionDetail(
  gint,
  extra_prom_gr,
  chr,
  bounds_offset = 15000,
  main = NULL
)
}
\arguments{
\item{gint}{genomic ranges object on a single chromosome}

\item{extra_prom_gr}{A Genomic Ranges object containing extra promoters that are disconnected from interactions}

\item{chr}{chromsome}

\item{bounds_offset}{Beyond the largest and smallest elements, how much extra space in bp should be plotted?}

\item{main}{character string for the title of the plot}
}
\description{
Creates a Gviz promoter-enhancer plot
}
\examples{
#See example at the start of the vignette
}
\keyword{GenomicInteractions}
\keyword{GenomicRanges}
\keyword{Gviz}
\keyword{metadata}
