% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGint.R
\name{readGint}
\alias{readGint}
\title{Create a GenomicInteractions object
with metadata}
\usage{
readGint(csvfile)
}
\arguments{
\item{csvfile}{Filename to be read}
}
\value{
gint a GenomicInteractions object with metadata
}
\description{
Creates a GenomicInteractions object from a CSV file
}
\examples{
readGint(system.file(
  "extdata","signficant_genes_genoscan_3D_knock.csv",
  package = "GS3DKViz"))
}
\keyword{GenomicInteractions}
\keyword{GenomicRanges}
\keyword{metadata}
