% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteraction.R
\name{plotInteraction}
\alias{plotInteraction}
\title{Plot a GenomicInteractions object}
\usage{
plotInteraction(gint, chr, bounds_offset = 15000, main = NULL)
}
\arguments{
\item{gint}{genomic ranges object on a single chromosome}

\item{chr}{chromsome}

\item{bounds_offset}{Beyond the largest and smallest elements, how much extra space in bp should be plotted?}
}
\description{
Creates a Gviz promoter-enhancer plot
}
\examples{
gs3dk_gint=readGint(system.file(
  "extdata","signficant_genes_genoscan_3D_knock.csv",
  package = "GS3DKViz"))
  filtered_gs3dk_gint=gs3dk_gint[(gs3dk_gint \%>\% as.data.frame() \%>\% janitor::clean_names() \%>\% dplyr::pull(trait))=="FEV1"]
  plotInteraction(gint = filtered_gs3dk_gint,
 chr=unique(janitor::clean_names(tibble::as_tibble(mcols(filtered_gs3dk_gint)))$chr),
 bounds_offset=1.5e4,
 main="Promoter–enhancer interactions, FEV1"
 )
}
\keyword{GenomicInteractions}
\keyword{GenomicRanges}
\keyword{Gviz}
\keyword{metadata}
